Feat: "추가 /login 리프레쉬 토큰 반환"

+++ src/main/java/repick/repickserver/global/config/JwtProperties.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Component;
 @Data
 public class JwtProperties {
     private String secret;
-    private int accessTokenExpirationTime;
-    private int refreshTokenExpirationTime;
+    private Long accessTokenExpirationTime;
+    private Long refreshTokenExpirationTime;
     private String authoritiesKey;
 }
